Allow setting Vulkan headers include path in cmake
Moreover, add an option to fail the build instead of downloading sources.
This allows the CTS to re-use the Vulkan headers already fetched by other
components, which also means aligning with the same Vulkan headers version.

With the new option FAIL_IF_NO_HEADERS, the CTS can enforce re-using its
own headers, and if they are not present or their version is not compatible
with the one required by vulkan-video-samples, it will fail the build
instead of downloading new ones.