EncoderConfigH264: use ADAPTIVE_TRANSFORM_AUTOSELECT
In order to use the baseline or main profile,
transform_8x8_mode_flag should not be set to true.
Use the auto select which is more appropriate here.

In case ADAPTIVE_TRANSFORM_ENABLE is used, disable it
and warn user that this option can not be used with
baseline and main profile.