FFmpegDemuxer: Strip higher bits of profile
H.264 uses the higher bits of this for constraint_set_flags. When these
extended values get passed to Vulkan as stdProfileIdc, the driver may
indicate they do not support such "bogus" values.

The only lavc codec that uses the higher bits is H.264, so it is safe to
keep only the lower bits.