VkEncoderConfigH264: improve level limits table and related code
The current formatting of the level limits table makes it hard to read
and compare against the source for the values within (i.e. table A-1 of
the H.264 specification). Further, we have a "prog" column in this table
whose entries have been taken from table A-4 of the H.264 specification
and specify the value to be used for the SPS frame_mbs_only_flag syntax
element for a given level, but these values are not used anywhere in the
code as interlaced encoding is not supported (and therefore
frame_mbs_only_flag = 1 always).

This change adds some whitespace to align the columns of this table and
make it more readable, while removing the unused "prog" column.

When editing the table, it was noticed that no entries were present for
levels 6.0, 6.1, 6.2 defined in table A-1. These entries have been added
now, and they allow removing some workarounds where the level was
previously artificially limited to level 5.2 as the maximum supported
level.

Finally, this change also fixes a bug in the level determination code in
EncoderConfigH264::DetermineLevel(). This code uses a fixed multiplier
equal to 1200 for the MaxBR and MaxCPB entries in the level limits
table, taken from table A-1 of the H.264 specification, but this
multiplier is to be used only for the Baseline, Constrained Baseline,
Main and Extended profiles. Profiles greater than these use a larger
multiplier, as specified in table A-2 of the H.264 specification. Due to
the use of a smaller multiplier, this code selects a greater level than
is strictly necessary in some situations, based on the input settings.

This commit updates this code to use a multiplier dependent on the codec
profile.