encode: remove "prog" column from H.264 level limits table
The values in this column have been taken from table A-4 of the H.264
specification and specify the value to be used for the SPS
frame_mbs_only_flag syntax element for a given level. These values are
not used anywhere in the code as interlaced encoding is not supported
(and therefore frame_mbs_only_flag = 1 always).

Signed-off-by: Srinath Kumarapuram <skr@nvidia.com>